Revert "`pj-rehearse`: refactor in place to allow for use in prow plugin"
