+++ test/integration/pj-rehearse.sh
@@ -43,9 +43,6 @@ export JOB_SPEC='{"type":"presubmit","job":"pull-ci-openshift-release-master-reh
 actual="${workdir}/rehearsals.yaml"
 os::cmd::expect_success "pj-rehearse --dry-run=true --candidate-path ${repo} --rehearsal-limit 20 > ${actual}"
 os::integration::sanitize_prowjob_yaml ${actual}
-# Substitute the SHA in the job names to a known SHA for comparison
-sed -i -E -e "s/-${candidate_sha}-/-4de8ab7c20656998264a2593116288f5eb070b32-/g" ${actual}
-
 os::integration::compare "${actual}" "${suite_dir}/expected.yaml"
 
 os::test::junit::declare_suite_end
